dom/file/ipc/tests/test_ipcBlob_fileReaderSync.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test IPCBlob and FileReaderSync</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script type="text/javascript">

function workerScript() {
  onmessage = function(event) {
    let readerMemoryBlob = new FileReaderSync();
    let status = readerMemoryBlob.readAsText(new Blob(['hello world'])) == 'hello world';

    let readerIPCBlob = new FileReaderSync();
    postMessage({ blob: event.data, data: readerIPCBlob.readAsText(event.data), status });
  }
}

let workerUrl = URL.createObjectURL(new Blob(["(", workerScript.toSource(), ")()"]));
let worker = new Worker(workerUrl);
worker.onmessage = event => {
  let fr = new FileReader();
  fr.readAsText(event.data.blob);
  fr.onload = () => {
    is(event.data.data, fr.result, "The file has been read");
    ok(event.data.status, "FileReaderSync with memory blob still works");
    SimpleTest.finish();
  }
};

let url = SimpleTest.getTestFileURL("script_file.js");
let script = SpecialPowers.loadChromeScript(url);
script.addMessageListener("file.opened", message => {
  worker.postMessage(message.file);
});

script.sendAsyncMessage("file.open");

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
